requirements.txt
Django==4.2
gunicorn==20.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
